pyproject.toml
[project]
name = "mcp-formlabs"
version = "0.1.0"
description = "MCP server for Formlabs PreForm Local API integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "trimesh>=4.0.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-telegram-bot>=21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mcp-formlabs = "mcp_formlabs.server:main"
mcp-formlabs-auth = "mcp_formlabs.auth_server:main"
mcp-formlabs-telegram = "mcp_formlabs.telegram_bot:main"
